+++ libsensors/lsm330dlc_acceleration.c
@@ -31,7 +31,7 @@
 #include "lsm330dlc_accel.h"
 
 struct lsm330dlc_acceleration_data {
-	long int delay;
+	int64_t delay;
 	int device_fd;
 	int uinput_fd;
 
@@ -47,7 +47,7 @@ void *lsm330dlc_acceleration_thread(void *thread_data)
 	struct input_event event;
 	struct timeval time;
 	struct lsm330dlc_acc acceleration_data;
-	long long int before, after;
+	int64_t before, after;
 	int diff;
 	int device_fd;
 	int uinput_fd;
@@ -285,14 +285,14 @@ int lsm330dlc_acceleration_deactivate(struct smdk4x12_sensors_handlers *handlers
 	return 0;
 }
 
-int lsm330dlc_acceleration_set_delay(struct smdk4x12_sensors_handlers *handlers, long int delay)
+int lsm330dlc_acceleration_set_delay(struct smdk4x12_sensors_handlers *handlers, int64_t delay)
 {
 	struct lsm330dlc_acceleration_data *data;
 	int64_t d;
 	int device_fd;
 	int rc;
 
-	ALOGD("%s(%p, %ld)", __func__, handlers, delay);
+	ALOGD("%s(%p, %" PRId64 ")", __func__, handlers, delay);
 
 	if (handlers == NULL || handlers->data == NULL)
 		return -EINVAL;
